% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractContextsSvLinx.R
\name{annotateComplexSVs}
\alias{annotateComplexSVs}
\title{Assign COMPLEX SV subtypes}
\usage{
annotateComplexSVs(df, verbose = F)
}
\arguments{
\item{df}{A dataframe of the *vis_sv_data.tsv file}

\item{verbose}{Show progress?}
}
\value{
A data.frame or character vector
}
\description{
Assign COMPLEX SV subtypes
}
