% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentAnalysis.R
\name{calcCancerTypeEnrichment}
\alias{calcCancerTypeEnrichment}
\title{Perform an enrichment (fisher test) analysis across multiple classes}
\usage{
calcCancerTypeEnrichment(
  df,
  feature.colname,
  class.colname = "cancer_type",
  v.classes,
  verbose = F
)
}
\arguments{
\item{df}{An input dataframe}

\item{feature.colname}{The colname of the feature (e.g. 'gene')}

\item{class.colname}{The colname of the classes (default: 'cancer_type')}

\item{v.classes}{A vector of classes from the cohort. Used to determine the
class background frequencies}

\item{verbose}{Show progress messages?}
}
\value{
A dataframe containing the fisher test data
}
\description{
Perform an enrichment (fisher test) analysis across multiple classes
}
\examples{
## Load data
metadata <- read.delim('/Users/lnguyen/hpc/cuppen/projects/P0013_WGS_patterns_Diagn//CUPs_classifier/processed/metadata/cancer_type_labels.txt')
metadata <- metadata[metadata$cohort=='HMF',]
driver_catalog <- read.delim('/Users/lnguyen/hpc/cuppen/projects/P0013_WGS_patterns_Diagn//CUPs_classifier/processed/cuplr/hmfSvFeatureExtractor/scripts/select_drivers//driver.catalog.merged.txt.gz')
driver_catalog$cancer_type <- metadata[match(driver_catalog$sample, metadata$sample),'cancer_type']
amp_catalog <- driver_catalog[driver_catalog$driver=='AMP',]

##
enrichment <- calcCancerTypeEnrichment(
  amp_catalog, feature.colname='gene',
  v.classes=metadata$cancer_type
)
}
