% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractContextsSvLinx.R
\name{countDelDupByLen}
\alias{countDelDupByLen}
\title{Count DELs and DUPs stratified by length}
\usage{
countDelDupByLen(
  linx.svs,
  bin.breaks = c(0, 300, 10^(3:7), Inf),
  sample.id.colname = NULL,
  output = "contexts"
)
}
\arguments{
\item{linx.svs}{A dataframe of a *linx.vis_sv_data.tsv file}

\item{bin.breaks}{An integer vector specifying the length bin breaks}

\item{sample.id.colname}{A string specifying the column name used by indicating the sample name.
If unspecified, all SVs are assumed to be from the same sample.}

\item{output}{'contexts' returns a matrix of the counts of each context. 'raw' returns a
dataframe with annotations for each SV (i.e. row).}
}
\value{
An integer vector
}
\description{
Count DELs and DUPs stratified by length
}
