% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countRepElemInsertions.R
\name{countRepElemInsertions}
\alias{countRepElemInsertions}
\title{Count number of insertions of repeat elements}
\usage{
countRepElemInsertions(
  vcf.file,
  rep.elem.whitelist.path = REP_ELEM_WHITELIST,
  tag.features = T
)
}
\arguments{
\item{vcf.file}{Path to vcf file}

\item{rep.elem.whitelist.path}{Path to txt file with list of selected repeat elements}
}
\value{
An integer vector
}
\description{
Count number of insertions of repeat elements
}
