% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractContextsSvLinx.R
\name{countSimpleSvEvents}
\alias{countSimpleSvEvents}
\title{Count DELs and DUPs stratified by length}
\usage{
countSimpleSvEvents(linx.svs, bin.breaks = c(0, 10^(3:7), Inf))
}
\arguments{
\item{linx.svs}{A dataframe of a *linx.vis_sv_data.tsv file}

\item{bin.breaks}{An integer vector specifying the length bin breaks}
}
\value{
An integer vector
}
\description{
Count DELs and DUPs stratified by length
}
