% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectKataegis.R
\name{detectKataegis}
\alias{detectKataegis}
\title{Detect kataegis from SNVs}
\usage{
detectKataegis(
  vcf.file = NULL,
  df = NULL,
  sample.name = NULL,
  vcf.filter = "PASS",
  keep.chroms = c(1:22, "X"),
  ref.genome = mutSigExtractor::DEFAULT_GENOME,
  min.muts = 6,
  imd.thres = 1000,
  cp = 1e-06,
  max.imd.quantile = 1,
  output.type = "count",
  verbose = T,
  ...
)
}
\arguments{
\item{vcf.file}{Path to the vcf file containing SNVs}

\item{df}{A bed file-like dataframe with the columns: chrom, pos, ref, alt}

\item{vcf.filter}{Which variants to filter for? Default is 'PASS'. Only applies if vcf.file is
specified}

\item{keep.chroms}{Which chromosomes to keep? Default is c(1:22,'X')}

\item{min.muts}{Min number of SNVs in a PCF segment for the segment to be considered a kataegis
focus}

\item{imd.thres}{Only PCF segments with mean intermutation distance less that this value will be
considered kataegis foci}

\item{cp}{rpart.control complexity parameter. Higher values increase overfitting of PCF}

\item{max.imd.quantile}{The intermutation distance at this quantile is calculated. If
max.imd.quantile < imd.thres, imd.thres will be set to max.imd.quantile}

\item{output.type}{'count': number of kataegis foci. 'kat_segments': simplified data of kataegis
segments (chrom and pos refer to the first variant in the kataegis focus). 'imd_kat': data from
all kataegis associated mutations, 'imd_all': raw data. Note that the last variant from each
chromosome is not reported (when output.type!='count')}

\item{verbose}{}

\item{...}{}

\item{merge.consecutive}{Some vcfs report MNVs as consecutive variants. For these vcfs, such rows
need to be merged into one row for proper function of downstream mutSigExtractor functions.}
}
\value{
An integer or dataframe
}
\description{
Detects kataegis from SNVs provided from a vcf file or dataframe. A piecewise
constant fit (using rpart (R in built regression tree algorithm)) is performed on the
intermutation distances. By default, segments with min.muts>=6 and imd.thres<=1000 are considered
kataegis foci. These default values were chosen according to d'Antonio et al Cell Rep. 2016
(https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4972030/)
}
