% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractContextsSmnvIndel.R
\name{extractContextsSmnvIndel}
\alias{extractContextsSmnvIndel}
\title{Extract SNV/indel contexts/signatures}
\usage{
extractContextsSmnvIndel(
  vcf.file = NULL,
  df = NULL,
  sample.name = NULL,
  vcf.filter = "PASS",
  keep.chroms = c(1:22, "X"),
  ref.genome = mutSigExtractor::DEFAULT_GENOME,
  clonal.variants.only = T,
  as.matrix = T,
  verbose = F,
  ...
)
}
\arguments{
\item{vcf.file}{Path to the vcf file}

\item{df}{A dataframe containing the columns: chrom, pos, ref, alt. Alternative input option to
vcf.file}

\item{sample.name}{If a character is provided, the header for the output
matrix will be named to this. If none is provided, the basename of the vcf
file will be used.}

\item{vcf.filter}{A character or character vector to specifying which variants to keep,
corresponding to the values in the vcf FILTER column}

\item{keep.chroms}{A character vector specifying which chromosomes to keep (chromosome names
should be in the style of the vcf). To keep autosomal and sex chromosomes for example use:
keep.chroms=c(1:22,'X','Y')}

\item{ref.genome}{A BSgenome reference genome. Default is BSgenome.Hsapiens.UCSC.hg19. If another
reference genome is indicated, it will also need to be installed.}

\item{clonal.variants.only}{If TRUE, only variants with subclonal likelihood (from GRIDSS) < 0.8
will be selected}

\item{as.matrix}{Return a matrix or vector?}

\item{verbose}{Print progress messages?}

\item{...}{Other arguments that can be passed to mutSigExtractor::variantsFromVcf()}
}
\value{
An matrix or integer vector
}
\description{
Extract SNV/indel contexts/signatures
}
