% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractContextsSvLinx.R
\name{extractContextsSvLinx}
\alias{extractContextsSvLinx}
\title{Extract SV contexts from LINX}
\usage{
extractContextsSvLinx(vis.sv.data = NULL)
}
\arguments{
\item{vis.sv.data}{Path to the *vis_sv_data.tsv file or a dataframe of this file}
}
\value{
An integer vector of counts
}
\description{
Extract SV contexts from LINX
}
