% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeaturesCuplr.R
\name{extractFeaturesCuplr}
\alias{extractFeaturesCuplr}
\title{Extract SV features from PURPLE and LINX output}
\usage{
extractFeaturesCuplr(
  purple.cnv.path,
  purple.purity.path,
  linx.fusion.path,
  linx.viral.inserts.path,
  driver.catalog.path,
  out.path = NULL,
  sample.name = NULL,
  verbose = 1
)
}
\arguments{
\item{purple.cnv.path}{Path to purple cnv file}

\item{purple.purity.path}{Path to purple.purity.tsv file}

\item{linx.fusion.path}{Path to LINX txt file containing fusion data}

\item{linx.viral.inserts.path}{Path to txt file with the columns: SampleId,
SvId, VirusId, VirusName}

\item{driver.catalog.path}{Path to LINX driver catalog txt file}

\item{out.path}{If provided, will write output to path as a 1 column dataframe with the
sample.name as the column name and feature names as the row names}

\item{sample.name}{See out.path}

\item{verbose}{Show messages?}
}
\value{
A 1-row data.frame
}
\description{
Extract SV features from PURPLE and LINX output
}
