% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRmd.R
\name{extractRmd}
\alias{extractRmd}
\title{Calculated regional mutation density (rmd)}
\usage{
extractRmd(
  vcf.file = NULL,
  df = NULL,
  sample.name = NULL,
  genome.bins = RMD_BINS_PATH,
  as.matrix = T,
  vcf.filter = "PASS",
  clonal.variants.only = F,
  mut.types = "snv",
  verbose = F,
  ...
)
}
\arguments{
\item{vcf.file}{Path to the vcf file}

\item{df}{A dataframe containing the columns: chrom, pos, ref, alt. Alternative input option to
vcf.file}

\item{sample.name}{If a character is provided, the header for the output
matrix will be named to this. If none is provided, the basename of the vcf
file will be used.}

\item{genome.bins}{A dataframe containing predetermined bins. Must contain the columns: chrom,
start, end, bin_name. Alternatively, if a path to a txt (as a string) is provided, the txt will
be loaded.}

\item{as.matrix}{Return a matrix or vector?}

\item{vcf.filter}{A character or character vector to specifying which variants to keep,
corresponding to the values in the vcf FILTER column}

\item{clonal.variants.only}{If TRUE, only variants with subclonal likelihood (from GRIDSS) < 0.8
will be selected}

\item{mut.types}{A character vector containing one or more of the following values: 'snv', 'dbs',
'indel', or 'other'. Default is 'snv'. Only these mut types will be taken into account when
calculating the number of mutations per RMD bin}

\item{verbose}{Print progress messages?}

\item{...}{Other arguments that can be passed to mutSigExtractor::variantsFromVcf()}
}
\value{
A matrix or vector
}
\description{
Calculated regional mutation density (rmd)
}
