% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkFeatureMatrix.R
\name{fillMissingSamplesInMatrix}
\alias{fillMissingSamplesInMatrix}
\title{Fill a matrix with a specified value for missing samples}
\usage{
fillMissingSamplesInMatrix(m, sample.names, fill.value = NA)
}
\arguments{
\item{m}{An input matrix}

\item{sample.names}{A character vector of all samples}

\item{fill.value}{What value to fill the missing sample rows (default is NA)}
}
\value{
A matrix
}
\description{
Fill a matrix with a specified value for missing samples
}
