% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrivers.R
\name{getDrivers}
\alias{getDrivers}
\title{Make gene driver features from gene diplotypes table}
\usage{
getDrivers(
  gene.diplotypes.path = NULL,
  gene.diplotypes = NULL,
  gene.amp.whitelist.path = GENE_AMP_WHITELIST,
  gene.def.whitelist.path = GENE_DEF_WHITELIST
)
}
\arguments{
\item{gene.diplotypes.path}{Path to gene_diplotypes.txt.gz (output from geneDriverAnnotator)}

\item{gene.diplotypes}{Alternatively the gene diplotypes dataframe can be provided as input}

\item{gene.amp.whitelist.path}{Path to txt file with list of selected amplified genes}

\item{gene.def.whitelist.path}{Path to txt file with list of selected Deficient genes}
}
\value{
A list: list(amp=character(), def=character())
}
\description{
Make gene driver features from gene diplotypes table
}
