% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneDriverEvents.R
\name{getGeneDriverEvents}
\alias{getGeneDriverEvents}
\title{Determine the presence of gene mono/biallelic hits, deep deletions, and amplifications}
\usage{
getGeneDriverEvents(
  linx.drivers,
  whitelist.path = GENE_DRIVER_WHITELIST,
  min.dnds.likelihood = 0.9,
  na.output = T
)
}
\arguments{
\item{linx.drivers}{Path to the LINX driver catalog txt file, or a dataframe of the file}

\item{whitelist.path}{Path to the gene whitelist txt file}

\item{min.dnds.likelihood}{Min DNDS likelihood for to consider a monoallelic hit as impactful}

\item{na.output}{If TRUE, and if \code{linx.drivers} is NA, then a vector of FALSE values will be
returned}
}
\value{
A named logical vector indicating which gene driver events are present in a sample
}
\description{
Determine the presence of gene mono/biallelic hits, deep deletions, and amplifications
}
