% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneFusions.R
\name{getGeneFusions}
\alias{getGeneFusions}
\title{Determine the presence of gene fusions}
\usage{
getGeneFusions(linx.fusions, whitelist.path = GENE_FUSION_WHITELIST)
}
\arguments{
\item{linx.fusions}{Path to the LINX fusion txt file, or a dataframe of the file}

\item{whitelist.path}{Path to gene fusion whitelist txt file}
}
\value{
A named logical vector indicating which gene fusions are present in a sample
}
\description{
Determine the presence of gene fusions
}
