% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getViralInsertions.R
\name{getViralInsertions}
\alias{getViralInsertions}
\title{Determines the presence of viral inserts in a sample}
\usage{
getViralInsertions(viral.inserts)
}
\arguments{
\item{viral.inserts}{Path to the LINX viral inserts txt file, or a dataframe of the file}
}
\value{
A named logical vector indicating which viruses are present in a sample
}
\description{
Determines the presence of viral inserts in a sample
}
