% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.split.R
\name{groupFeaturesByTag}
\alias{groupFeaturesByTag}
\title{Group a vector of feature names by their feature tags}
\usage{
groupFeaturesByTag(x, rm.tags = F)
}
\arguments{
\item{x}{A character vector in the form 'feature_tag.feature_name'}

\item{rm.tags}{Remove feature group tags?}
}
\value{
A list of feature vectors in for each tag
}
\description{
Group a vector of feature names by their feature tags
}
