% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeUtils.R
\name{linearizeChromPos}
\alias{linearizeChromPos}
\title{Converts chrom/pos into linear coordinates}
\usage{
linearizeChromPos(
  chrom = NULL,
  pos = NULL,
  df = NULL,
  chrom.lengths = CHROM_LENGTHS
)
}
\arguments{
\item{chrom}{A character vector indicating the chromosome}

\item{pos}{An integer vector indicating the position}

\item{df}{Alternative input to chrom and pos. A dataframe with chrom in 1st col and pos in 2nd col}

\item{chrom.lengths}{A named vector of chromosome lengths (default: hg19)}
}
\value{
An integer vector of linear coordinates
}
\description{
Converts chrom/pos into linear coordinates
}
