% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeaturesCuplr.R
\name{mkFeatureMatrix}
\alias{mkFeatureMatrix}
\title{Execute an extraction function from hmfSvFeatureExtractor over a list of files}
\usage{
mkFeatureMatrix(
  file.paths,
  sample.names,
  func,
  func.args = list(),
  verbose = T
)
}
\arguments{
\item{file.paths}{A vector of file paths that will be passed to \code{func}}

\item{sample.names}{A vector of sample names}

\item{func}{A function to execute}

\item{func.args}{Other args for \code{func}}

\item{verbose}{Show progress messages?}
}
\value{
A matrix
}
\description{
Execute an extraction function from hmfSvFeatureExtractor over a list of files
}
\examples{
manifest <- read.delim('/Users/lnguyen/hpc/cuppen/projects/P0013_WGS_patterns_Diagn/CUPs_classifier/processed/cuplr/hmfSvFeatureExtractor/scripts/mk_manifest/manifest.txt.gz')
if(dir.exists('/Users/lnguyen/')){
   counter <- 0
   manifest <- as.data.frame(lapply(manifest, function(i){
      counter <<- counter + 1
      if(counter==1){ return(i) }
      paste0('/Users/lnguyen/',i)
   }))
}

file_paths <- manifest$linx.fusion[1:10]
sample_names <- sapply(strsplit(basename(file_paths),'[.]'),`[[`,1)
m <- mkFeatureMatrix(
   file.paths=file_paths,
   sample.names=sample_names,
   func=getGeneFusions
)

write.table(
   m,
   gzfile('/Users/lnguyen/hpc/cuppen/projects/P0013_WGS_patterns_Diagn/CUPs_classifier/processed/cuplr/hmfSvFeatureExtractor/scripts/mk_matrices/fusions/m_fusions.txt.gz'),
   sep='\t',quote=F
)
}
