% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractContextsSvLinx.R
\name{mkIntervalLookup}
\alias{mkIntervalLookup}
\title{Make interval lookup table}
\usage{
mkIntervalLookup(bins = c(0, 10^c(3:7)))
}
\arguments{
\item{bins}{A numeric vector specifying the interval breaks}
}
\value{
A dataframe
}
\description{
Make interval lookup table
}
