% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentAnalysis.R
\name{plotCancerTypeEnrichment}
\alias{plotCancerTypeEnrichment}
\title{Plot enrichment analysis}
\usage{
plotCancerTypeEnrichment(
  df,
  keep.y.order = F,
  invert.y = T,
  show.top.n.features = NULL,
  show.qvalues = F
)
}
\arguments{
\item{df}{The output from calcCancerTypeEnrichment()}

\item{keep.y.order}{Force the original order of the features?}

\item{invert.y}{Invert the y-axis?}

\item{show.top.n.features}{Show only n features (subset y axis)}

\item{show.qvalues}{Show q values instead of p values?}
}
\value{
A ggplot object
}
\description{
Plot enrichment analysis
}
