% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeaturesCuplr.R
\name{readFeaturesCuplr}
\alias{readFeaturesCuplr}
\title{Read and format features for CUPLR}
\usage{
readFeaturesCuplr(x)
}
\arguments{
\item{x}{Path to the features.txt.gz file, or the output dataframe from \code{extractFeaturesCuplr()}}
}
\value{
A 1-row dataframe
}
\description{
Read features from a features.txt.gz file and assign factor levels for categorical
features
}
