% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcChromArmPloidies.main.R
\name{calcChromArmCnChange}
\alias{calcChromArmCnChange}
\alias{calcChromArmCnChange.numeric}
\alias{calcChromArmCnChange.matrix}
\alias{calcChromArmCnChange.data.frame}
\title{Calculate change in chrom arm copy number compared to genome copy number}
\usage{
calcChromArmCnChange(x, ...)

\method{calcChromArmCnChange}{numeric}(x, direction = "both")

\method{calcChromArmCnChange}{matrix}(x, direction = "both")

\method{calcChromArmCnChange}{data.frame}(x, direction = "both")
}
\arguments{
\item{x}{Output from calcChromArmPloidies() as a matrix or vector}

\item{direction}{Can be'gain', 'loss' or 'both'}
}
\value{
A matrix or vector
}
\description{
Calculate change in chrom arm copy number compared to genome copy number
}
