% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractContextsSvLinx.R
\name{extractContextsSvLinx}
\alias{extractContextsSvLinx}
\title{Extract SV contexts from LINX}
\usage{
extractContextsSvLinx(
  vis.sv.data.path = NULL,
  df = NULL,
  len.bins.del = c(0, 10^c(3:7), Inf),
  len.bins.dup = len.bins.del,
  resolved.type.annotations.path = RESOLVED_TYPE_ANNOTATIONS,
  split.complex = F
)
}
\arguments{
\item{vis.sv.data.path}{Path to the *vis_sv_data.tsv file}

\item{df}{A dataframe of the *vis_sv_data.tsv file}

\item{len.bins.del}{A numeric vector specifying the interval breaks to bin DEL lengths}

\item{len.bins.dup}{A numeric vector specifying the interval breaks to bin DUP lengths}

\item{resolved.type.annotations.path}{Path to the txt file containing annotations for each LINX
ResolvedType}

\item{split.complex}{If TRUE, COMPLEX SVs will be split into foldbacks, hiSvLoad,
fewChrom_ploidyOscill, multiChrom_ploidyStable, ploidyUnstable, and other}
}
\value{
An integer vector of counts
}
\description{
Extract SV contexts from LINX
}
