% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeaturesCuplr.R
\name{extractFeaturesCuplr}
\alias{extractFeaturesCuplr}
\title{Extract input features for CUPLR}
\usage{
extractFeaturesCuplr(
  in.dir = NULL,
  input.paths = NULL,
  out.dir = NULL,
  clonal.variants.only = F,
  verbose = F
)
}
\arguments{
\item{in.dir}{A directory containing the input files. The required files are searched for with
the following regular expressions:
\itemize{
\item  purple.smnv:        '.purple.somatic.vcf.gz$'
\item  purple.cnv:         '.purple.cnv.somatic.tsv$'
\item  purple.purity:      '.purple.purity.tsv$'
\item  linx.drivers:       '.linx.driver.catalog.tsv$'
\item  linx.fusions:       '.linx.fusion.tsv$'
\item  linx.viral.inserts: '.linx.viral_inserts.tsv$'
\item  linx.vis.sv.data:   '.linx.vis_sv_data.tsv$'
}}

\item{input.paths}{A named list of the input paths. See \code{in.dir} for the required names}

\item{out.dir}{(Optional) A directory to write intermediate and final output files. If
unspecified, a dataframe of the features will be returned}

\item{clonal.variants.only}{If TRUE, only clonal variants will be used for the mutational
signature and RMD signature features}

\item{verbose}{Show progress messages? 0: No messages. 1: Messages from \code{extractFeaturesCuplr()}.
2: Messages from internal functions called by \code{extractFeaturesCuplr()}}
}
\value{
See \code{out.dir}
}
\description{
Extract input features for CUPLR
}
