% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeaturesCuplr.R
\name{extractFeaturesCuplr}
\alias{extractFeaturesCuplr}
\title{Extract features for CUPLR}
\usage{
extractFeaturesCuplr(
  germ.vcf.path,
  som.vcf.path,
  sv.vcf.path,
  purple.cnv.path,
  purple.purity.path,
  linx.fusion.path,
  linx.viral.inserts.path,
  linx.vis.sv.data.path,
  colname.translations = list(),
  out.dir,
  sample.name,
  verbose = 1
)
}
\arguments{
\item{germ.vcf.path}{Path to germline SNV/indel vcf}

\item{som.vcf.path}{Path to somatic SNV/indel vcf}

\item{sv.vcf.path}{Path to GRIDSS SV vcf}

\item{purple.cnv.path}{Path to purple cnv file}

\item{purple.purity.path}{Path to purple.purity.tsv file}

\item{linx.fusion.path}{Path to LINX txt file containing fusion data}

\item{linx.viral.inserts.path}{Path to txt file with the columns: SampleId,}

\item{linx.vis.sv.data.path}{Path to the LINX txt file containing the annotated SVs}

\item{out.dir}{out.dir Path to output dir}

\item{sample.name}{A character string indicating the sample name}

\item{verbose}{Show messages?}
}
\value{
A 1-row data.frame
}
\description{
Extract features for CUPLR
}
