% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeaturesCuplr.R
\name{extractFeaturesCuplr}
\alias{extractFeaturesCuplr}
\title{Extract features for CUPLR}
\usage{
extractFeaturesCuplr(
  germ.vcf.path,
  som.vcf.path,
  sv.vcf.path,
  purple.cnv.path,
  purple.purity.path,
  linx.fusion.path,
  linx.viral.inserts.path,
  linx.vis.sv.data.path,
  colname.translations = list(),
  out.dir,
  sample.name,
  return.features = FALSE,
  write.features = TRUE,
  rm.tmp.files = FALSE,
  verbose = 1
)
}
\arguments{
\item{germ.vcf.path}{Path to germline SNV/indel vcf}

\item{som.vcf.path}{Path to somatic SNV/indel vcf}

\item{sv.vcf.path}{Path to GRIDSS SV vcf}

\item{purple.cnv.path}{Path to purple cnv file}

\item{purple.purity.path}{Path to purple.purity.tsv file}

\item{linx.fusion.path}{Path to LINX txt file containing fusion data}

\item{linx.viral.inserts.path}{Path to txt file with the columns: SampleId,}

\item{linx.vis.sv.data.path}{Path to the LINX txt file containing the annotated SVs}

\item{colname.translations}{Used to translate input table colnames to colnames that are
compatible with CUPLR. This is provided as a list with one or more of the following names:
purple.cnv, linx.vis.sv.data, linx.viral.inserts, linx.fusion. Each object should be a named
character vector, where names correspond to the colnames required by CUPLR, and values correspond
to the colnames in the input tables}

\item{out.dir}{out.dir Path to output dir}

\item{sample.name}{A character string indicating the sample name. If NULL, will use the basename
of \code{som.vcf.path}}

\item{return.features}{If TRUE, will return the dataframe of features which can be used directly
in R}

\item{write.features}{If TRUE, will write the extracted features to features.txt.gz at \code{out.dir}}

\item{rm.tmp.files}{If TRUE, will remove temporary files from \code{out.dir}}

\item{verbose}{Show messages?}
}
\value{
A 1-row data.frame
}
\description{
Extract features for CUPLR
}
