% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRmd.R
\name{extractRmd}
\alias{extractRmd}
\title{Calculated regional mutation density (rmd)}
\usage{
extractRmd(
  vcf.file = NULL,
  df = NULL,
  sample.name = NULL,
  genome.bins = NULL,
  bin.size = 1e+06,
  as.matrix = T,
  vcf.filter = "PASS",
  clonal.variants.only = T,
  mut.types = "snv",
  verbose = F,
  ...
)
}
\arguments{
\item{vcf.file}{Path to the vcf file}

\item{sample.name}{If a character is provided, the header for the output
matrix will be named to this. If none is provided, the basename of the vcf
file will be used.}

\item{genome.bins}{A dataframe containing predetermined bins. Must contain the columns: chrom,
start, end, chrom_arm, start_linear, end_linear, bin_name}

\item{bin.size}{Alternatively if \code{genome.bins} is NULL, the user can specify the size of the
genomic bins}

\item{as.matrix}{Return a matrix or vector?}

\item{vcf.filter}{A character or character vector to specifying which variants to keep,
corresponding to the values in the vcf FILTER column}

\item{clonal.variants.only}{If TRUE, only variants with subclonal likelihood (from GRIDSS) < 0.8
will be selected}

\item{verbose}{Print progress messages?}

\item{...}{Other arguments that can be passed to mutSigExtractor::variantsFromVcf()}
}
\value{
A matrix or vector
}
\description{
Calculated regional mutation density (rmd)
}
