% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeUtils.R
\name{mkGenomeBins}
\alias{mkGenomeBins}
\title{Make a look up table of genomic bin intervals}
\usage{
mkGenomeBins(
  bin.size = 1e+06,
  keep.chroms = c(1:22, "X"),
  split.centro.bins = T,
  centro.pos = CENTRO_POS_HG19,
  chrom.lengths = CHROM_LENGTHS_HG19
)
}
\arguments{
\item{bin.size}{Size of each genome bin}

\item{keep.chroms}{A character vector specifying which chromosomes to keep (chromosome names
should be in the style of the vcf). To keep autosomal and sex chromosomes for example use:
keep.chroms=c(1:22,'X','Y')}

\item{split.centro.bins}{When bins overlap centromere positions, split these into 2 bins?}

\item{centro.pos}{A named integer vector of the centromere positions (default: hg19)}

\item{chrom.lengths}{A named vector of chromosome lengths (default: hg19)}
}
\value{
A dataframe containing the intervals of each bin (in chr:start:end and
linear_start:linear_end form)
}
\description{
Make a look up table of genomic bin intervals
}
