% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeUtils.R
\name{setGenome}
\alias{setGenome}
\alias{getCurrentGenome}
\title{Set the genome to hg19 or hg38}
\usage{
setGenome(genome.name, verbose = 2)

getCurrentGenome()
}
\arguments{
\item{genome.name}{Can be 'hg19' or 'hg38'}

\item{verbose}{Show details?}
}
\description{
\code{setGenome()} assigns several global variables in the \code{featureExtractor} environment
including BSGENOME, CENTRO_POS, CHROM_LENGTHS and RMD_BINS_PATH. \code{getCurrentGenome()} displays
the current genome and the associated assigned global variables.
}
